requirements.txt
# Core LangChain / LangGraph
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# ── LLM backends (install the one you use) ───────────────────────────────────
# Option A — Ollama (local Llama 3, recommended, no quota)
langchain-ollama>=0.1.0

# Option B — Gemini (needs GEMINI_API_KEY)
langchain-google-genai>=2.0.0

# Streamlit UI
streamlit>=1.35.0

# Utilities
rich>=13.7.0

# ── PHI De-identification — Microsoft Presidio (dual-layer with regex) ────────
presidio-analyzer>=2.2.354
presidio-anonymizer>=2.2.354
# spaCy NLP engine for Presidio (download model separately — see README)
spacy>=3.7.0
# After install run:  python -m spacy download en_core_web_lg
# Fallback model  :  python -m spacy download en_core_web_sm

# ── Semantic retrieval (SBERT + FAISS) ───────────────────────────────────────
# PubMedBERT-based SBERT encoder for clinical text
sentence-transformers>=2.7.0
# FAISS — CPU build (no GPU required)
faiss-cpu>=1.8.0
# scikit-learn — TF-IDF fallback encoder when torch is unavailable
scikit-learn>=1.4.0
